backend/pyproject.toml
[project]
name = "jaguardata-platform"
version = "0.1.0"
description = "JaguarData Platform — Data Contract Manager + Generative Agents Engine"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.0",
    "jinja2>=3.1",
    "boto3>=1.35",
    "mangum>=0.17",
    "python-jose[cryptography]>=3.3",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "moto[all]>=5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
